first version of installer, move runtime location, get rid of build warnings
installer mostly modelled on autocad-ballet (but is mostly much simpler,
well except for keyboard shortcuts it is)

installer isn't complete yet, extensive testing requried, codebase needs
to be adjusted to support selected range to revit versions

installer registers commands in Windows registry to avoid user prompts
on first load

includes uninstaller, multi-version support with auto-detection

keyboard shortcut management is limited due to Revit not generating
KeyboardShortcuts.xml on start - instructions for the user were added to installer

includes update mechanism which avoids restarting revit (as much as possible)
detects if files are locked, if not then it just overrides them with new ones
if they are locked it copies them to commands/bin/<year>/release for invokeaddincommand to use
it creates a new folder revit-ballet.update next to the old revit-ballet folder and updates revit-ballet.addin to point to it
then revit-ballet replaces revit-ballet folder contents with contents from revit-ballet.update and updates revit-ballet.addin to point to it
next time revit runs it deletes revit-ballet.update folder (after checking that it's contents matches current revit-ballet folder)

runtime location moved to %appdata%/revit-ballet/runtime

build warnings complained about using deprecated revit api, changed from ElementId.IntegerValue to ElementId.Value

all startup code separated into a new file

+++ commands/FilterFilledRegions.cs
@@ -75,7 +75,7 @@ public static class ExtendedElementDataHelper
             }
             else if (data.ContainsKey("Id") && data["Id"] is int id)
             {
-                element = elementDoc.GetElement(new ElementId(id));
+                element = elementDoc.GetElement(new ElementId((long)id));
             }
             
             if (element != null)
@@ -361,7 +361,7 @@ public class FilterFilledRegions : IExternalCommand
                 }
                 else if (fullData.TryGetValue("Id", out var intId) && intId is int id)
                 {
-                    regularIds.Add(new ElementId(id));
+                    regularIds.Add(new ElementId((long)id));
                 }
             }
